added basic map functionality with button to toggle the map
