+++ package.json
@@ -12,12 +12,14 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@mui/icons-material": "^5.15.15",
     "@mui/material": "^5.15.15",
+    "@react-google-maps/api": "^2.19.3",
     "axios": "^1.6.7",
     "bootstrap": "^5.3.3",
     "mui": "^0.0.1",
     "react": "^18.2.0",
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^18.2.0",
+    "react-google-maps": "^9.4.5",
     "react-ratings-declarative": "^3.4.1",
     "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
@@ -50,7 +52,7 @@
     ]
   },
   "devDependencies": {
-    "@types/react": "^18.2.64",
+    "@types/react": "^16.14.60",
     "@types/react-dom": "^18.2.21",
     "gh-pages": "^6.1.1"
   }
